requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
reportlab==4.2.0
python-dateutil==2.8.2
Werkzeug==3.0.1
